[desgin] header 컴포넌트분리
